STICKER WALL completed
Only visible to signed in users where they can create posts on the sticker wall with text and predefined stickers. The posts would be visible to all signed in users where they have the option to like it. There is also an option for deleting post by the original poster.

Also, css needed for
-Friends leaderboard
-Sticker wall
-The buttons in Friends list too
-Some buttons need to be repositioned